[test] apply the new test style to databend-meta
