+++ metasrv/src/api/mod.rs
@@ -15,12 +15,9 @@
 // The api module only used for internal communication, such as GRPC between cluster and the managed HTTP REST API.
 
 mod flight_server;
-mod http;
+pub mod http;
 mod http_service;
 pub mod rpc;
-
-#[cfg(test)]
-mod http_service_test;
 
 pub use flight_server::FlightServer;
 pub use http_service::HttpService;
