+++ metasrv/tests/it/api/flight/metasrv_flight_meta_api.rs
@@ -15,14 +15,15 @@
 use common_base::tokio;
 use common_meta_api::MetaApiTestSuite;
 use common_meta_flight::MetaFlightClient;
-use databend_meta::init_meta_ut;
+
+use crate::init_meta_ut;
 
 #[tokio::test(flavor = "multi_thread", worker_threads = 1)]
 async fn test_meta_api_database_create_get_drop() -> anyhow::Result<()> {
     let (_log_guards, ut_span) = init_meta_ut!();
     let _ent = ut_span.enter();
 
-    let (_tc, addr) = databend_meta::tests::start_metasrv().await?;
+    let (_tc, addr) = crate::tests::start_metasrv().await?;
 
     let client = MetaFlightClient::try_create(addr.as_str(), "root", "xxx").await?;
 
@@ -34,7 +35,7 @@ async fn test_meta_api_database_list() -> anyhow::Result<()> {
     let (_log_guards, ut_span) = init_meta_ut!();
     let _ent = ut_span.enter();
 
-    let (_tc, addr) = databend_meta::tests::start_metasrv().await?;
+    let (_tc, addr) = crate::tests::start_metasrv().await?;
 
     let client = MetaFlightClient::try_create(addr.as_str(), "root", "xxx").await?;
 
@@ -46,7 +47,7 @@ async fn test_meta_api_table_create_get_drop() -> anyhow::Result<()> {
     let (_log_guards, ut_span) = init_meta_ut!();
     let _ent = ut_span.enter();
 
-    let (_tc, addr) = databend_meta::tests::start_metasrv().await?;
+    let (_tc, addr) = crate::tests::start_metasrv().await?;
 
     let client = MetaFlightClient::try_create(addr.as_str(), "root", "xxx").await?;
 
@@ -58,7 +59,7 @@ async fn test_meta_api_table_list() -> anyhow::Result<()> {
     let (_log_guards, ut_span) = init_meta_ut!();
     let _ent = ut_span.enter();
 
-    let (_tc, addr) = databend_meta::tests::start_metasrv().await?;
+    let (_tc, addr) = crate::tests::start_metasrv().await?;
 
     let client = MetaFlightClient::try_create(addr.as_str(), "root", "xxx").await?;
 
@@ -73,7 +74,7 @@ async fn test_meta_api_table_list() -> anyhow::Result<()> {
 async fn test_meta_api_flight_get_database_meta_ddl_table() -> anyhow::Result<()> {
     let (_log_guards, ut_span) = init_meta_ut!();
     let _ent = ut_span.enter();
-    let (_tc, addr) = databend_meta::tests::start_metasrv().await?;
+    let (_tc, addr) = crate::tests::start_metasrv().await?;
     let client = MetaFlightClient::try_create(addr.as_str(), "root", "xxx").await?;
 
     let test_db = "db1";
@@ -151,7 +152,7 @@ async fn test_meta_api_flight_get_database_meta_ddl_table() -> anyhow::Result<()
 async fn test_meta_api_flight_get_database_meta_empty_db() -> anyhow::Result<()> {
     let (_log_guards, ut_span) = init_meta_ut!();
     let _ent = ut_span.enter();
-    let (_tc, addr) = databend_meta::tests::start_metasrv().await?;
+    let (_tc, addr) = crate::tests::start_metasrv().await?;
     let client = MetaFlightClient::try_create(addr.as_str(), "root", "xxx").await?;
 
     // Empty Database
@@ -165,7 +166,7 @@ async fn test_meta_api_flight_get_database_meta_empty_db() -> anyhow::Result<()>
 async fn test_meta_api_flight_get_database_meta_ddl_db() -> anyhow::Result<()> {
     let (_log_guards, ut_span) = init_meta_ut!();
     let _ent = ut_span.enter();
-    let (_tc, addr) = databend_meta::tests::start_metasrv().await?;
+    let (_tc, addr) = crate::tests::start_metasrv().await?;
     let client = MetaFlightClient::try_create(addr.as_str(), "root", "xxx").await?;
 
     // create-db operation will increases meta_version
