+++ metasrv/tests/it/api/mod.rs
@@ -12,6 +12,6 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-mod any_error_impl;
-
-pub use any_error_impl::AnyError;
+mod flight;
+mod http;
+mod http_service;
