+++ metasrv/tests/it/main.rs
@@ -1,4 +1,4 @@
-// Copyright 2020 Datafuse Labs.
+// Copyright 2021 Datafuse Labs.
 //
 // Licensed under the Apache License, Version 2.0 (the "License");
 // you may not use this file except in compliance with the License.
@@ -12,6 +12,9 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-mod any_error_impl;
-
-pub use any_error_impl::AnyError;
+mod any_error;
+mod api;
+mod configs;
+mod meta_service;
+mod store;
+mod tests;
