+++ metasrv/tests/it/meta_service/mod.rs
@@ -12,11 +12,5 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-#[macro_use]
-pub mod service;
-pub mod tls_constants;
-
-pub use service::assert_meta_connection;
-pub use service::next_port;
-pub use service::start_metasrv;
-pub use service::start_metasrv_with_context;
+mod meta_service_impl;
+mod raftmeta;
